[Upd] Moved all in tasks.py for celery
