+++ tooling/database.py
@@ -1,5 +1,4 @@
 import redis
-from django.conf import settings
 
 
-conn = redis.Redis(host=settings.REDIS_HOST, port=6379, db=0)
+conn = redis.Redis(host="localhost", port=6379, db=0)
